fixed bug when force unwinding coroutines
Because when coroutines are going to be destroyed, so we may not have
the correct current_coro in Processor. But all we know is all running
coroutines must be resumed by the Processor. So I added methods to
resume the coroutine right after the callback function is finished by
the main_coro in the Processor.