Processor/Scheduler overhaul
This commit introduce steal-half ring buffers for Processor and
a unbounded global work queue for the Scheduler.
This idea and the related algorithms stem from Go’s source code.
Furthermore this commit introduces many internal improvements for the
communication between Processor and Scheduler and multiple
improvements which accumulated over the time.